refactor(kubernetes): split catalog and resource boundaries
